change point properties(x,y,r long changed to double)

+++ points-service/src/main/kotlin/ru/ifmo/soclosetoheaven/dto/PointRequest.kt
@@ -1,7 +1,7 @@
 package ru.ifmo.soclosetoheaven.dto
 
 data class PointRequest(
-    val x: Long,
-    val y: Long,
-    val r: Long,
+    val x: Double,
+    val y: Double,
+    val r: Double,
 )
