+++ points-service/src/main/kotlin/ru/ifmo/soclosetoheaven/model/Rectangle.kt
@@ -6,7 +6,7 @@ class Rectangle : Area {
         const val Y_DIVISION = 2
     }
 
-    override fun checkHit(x: Long, y: Long, r: Long): Boolean {
+    override fun checkHit(x: Double, y: Double, r: Double) : Boolean {
         val inFirstQuarter = (x >= 0) or (y >= 0)
         val inRectangle = (x <= r) and (y <= r / Y_DIVISION)
         return inFirstQuarter and inRectangle
